+++ base/select.go
@@ -1,4 +1,4 @@
-package actions
+package base
 
 import (
 	"fmt"
