+++ base/select_test.go
@@ -1,12 +1,10 @@
-package actions
+package base
 
 import (
 	"testing"
 
 	"github.com/qri-io/cafs"
 	"github.com/qri-io/dataset/dstest"
-	"github.com/qri-io/qri/config"
-	"github.com/qri-io/qri/p2p"
 	"github.com/qri-io/qri/repo"
 	"github.com/qri-io/qri/repo/profile"
 	regmock "github.com/qri-io/registry/regserver/mock"
@@ -20,10 +18,6 @@ func TestSelect(t *testing.T) {
 	if err != nil {
 		t.Fatal(err.Error())
 	}
-	node, err := p2p.NewQriNode(mr, config.DefaultP2PForTesting())
-	if err != nil {
-		t.Fatal(err.Error())
-	}
 
 	tc, err := dstest.NewTestCaseFromDir(testdataPath("cities"))
 	if err != nil {
@@ -31,10 +25,9 @@ func TestSelect(t *testing.T) {
 		return
 	}
 
-	ref, _, err := SaveDataset(node, tc.Name, tc.Input, tc.BodyFile(), nil, false, true)
+	ref, err := CreateDataset(mr, tc.Name, tc.Input, tc.BodyFile(), false)
 	if err != nil {
 		t.Fatal(err.Error())
-
 	}
 
 	if _, err := Select(mr, ref, "commit"); err != nil {
